Use BufferBuilder and UnsafeAppend to accelerate writes.

Before (prior to ARROW-6381, which came up while investigating this):

```
----------------------------------------------------------------------------------
Benchmark                                           Time           CPU Iterations
----------------------------------------------------------------------------------
BM_ArrowBinaryPlain/EncodeArrow/262144        6111690 ns    6109829 ns        465    246.06MB/s
BM_ArrowBinaryPlain/EncodeArrow/1048576      30470849 ns   30451048 ns         85   197.296MB/s
BM_ArrowBinaryPlain/EncodeLowLevel/262144     5352838 ns    5352679 ns        514   280.866MB/s
BM_ArrowBinaryPlain/EncodeLowLevel/1048576   29736017 ns   29735036 ns         94   202.047MB/s
```

After

```
----------------------------------------------------------------------------------
Benchmark                                           Time           CPU Iterations
----------------------------------------------------------------------------------
BM_ArrowBinaryPlain/EncodeArrow/262144        2020914 ns    2020905 ns       1000   743.918MB/s
BM_ArrowBinaryPlain/EncodeArrow/1048576      11596223 ns   11596094 ns        242   518.096MB/s
BM_ArrowBinaryPlain/EncodeLowLevel/262144     2740316 ns    2740256 ns       1021    548.63MB/s
BM_ArrowBinaryPlain/EncodeLowLevel/1048576   17562138 ns   17560763 ns        157    342.12MB/s
```

Dictionary encoding perf is not really affected by this work, so this will mostly affect that case where we fall back to PLAIN encoding when the dictionary grows large.