@wesm 2019-08-29
Python benchmarks show an end-to-end improvement of about 10%

Before this and ARROW-6831:

```
[ 12.50%] ··· ...rquetWriteBinary.time_convert_pandas_and_write_binary_table            2.16±0s
[ 25.00%] ··· parquet.ParquetWriteBinary.time_write_binary_table                        1.25±0s
[ 37.50%] ··· ...et.ParquetWriteBinary.time_write_binary_table_no_dictionary            1.24±0s
[ 50.00%] ··· ...uet.ParquetWriteBinary.time_write_binary_table_uncompressed            927±0ms
```

After

```
[ 12.50%] ··· ...rquetWriteBinary.time_convert_pandas_and_write_binary_table            1.96±0s
[ 25.00%] ··· parquet.ParquetWriteBinary.time_write_binary_table                        1.10±0s
[ 37.50%] ··· ...et.ParquetWriteBinary.time_write_binary_table_no_dictionary            1.08±0s
[ 50.00%] ··· ...uet.ParquetWriteBinary.time_write_binary_table_uncompressed            772±0ms
```